+++ src/lib.rs
@@ -1,8 +1,7 @@
 mod access_log_column_vecs;
 mod bridge;
 pub mod parsers;
-mod settings;
+pub mod settings;
 
 pub use access_log_column_vecs::AccessLogColumnVecs;
 pub use bridge::Bridge;
-pub use settings::Settings;